Revert "separate the header files"
