+++ Jongeun/BlackJack/BlackJack.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34723.18
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "BlackJack", "BlackJack\BlackJack.vcxproj", "{4EAF5DB2-67AB-49D2-8D0A-2B83E769276C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4EAF5DB2-67AB-49D2-8D0A-2B83E769276C}.Debug|x64.ActiveCfg = Debug|x64
+		{4EAF5DB2-67AB-49D2-8D0A-2B83E769276C}.Debug|x64.Build.0 = Debug|x64
+		{4EAF5DB2-67AB-49D2-8D0A-2B83E769276C}.Debug|x86.ActiveCfg = Debug|Win32
+		{4EAF5DB2-67AB-49D2-8D0A-2B83E769276C}.Debug|x86.Build.0 = Debug|Win32
+		{4EAF5DB2-67AB-49D2-8D0A-2B83E769276C}.Release|x64.ActiveCfg = Release|x64
+		{4EAF5DB2-67AB-49D2-8D0A-2B83E769276C}.Release|x64.Build.0 = Release|x64
+		{4EAF5DB2-67AB-49D2-8D0A-2B83E769276C}.Release|x86.ActiveCfg = Release|Win32
+		{4EAF5DB2-67AB-49D2-8D0A-2B83E769276C}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {229720DE-DB0B-40C3-9FA9-0B86A102A8B9}
+	EndGlobalSection
+EndGlobal
