+++ Jongeun/BlackJack/BlackJack/BlackJack.cpp
@@ -20,11 +20,115 @@ TODO: Implemented List
 #include <cassert>
 #include <iostream>
 #include "Random.h"
-#include "Card.h"
-#include "Deck.h"
-#include "Player.h"
 
+struct Card
+{
+    enum Rank
+    {
+        rank_ace,
+        rank_2,
+        rank_3,
+        rank_4,
+        rank_5,
+        rank_6,
+        rank_7,
+        rank_8,
+        rank_9,
+        rank_10,
+        rank_jack,
+        rank_queen,
+        rank_king,
 
+        max_ranks
+    };
+
+    enum Suit
+    {
+        suit_club,
+        suit_diamond,
+        suit_heart,
+        suit_spade,
+
+        max_suits
+    };
+
+    static constexpr std::array allRanks{rank_ace, rank_2, rank_3, rank_4, rank_5, rank_6, rank_7, rank_8, rank_9, rank_10, rank_jack, rank_queen, rank_king};
+    static constexpr std::array allSuits{suit_club, suit_diamond, suit_heart, suit_spade};
+
+    Rank rank{};
+    Suit suit{};
+
+    friend std::ostream &operator<<(std::ostream &out, const Card &card)
+    {
+        static constexpr std::array ranks{'A', '2', '3', '4', '5', '6', '7', '8', '9', 'T', 'J', 'Q', 'K'};
+        static constexpr std::array suits{'C', 'D', 'H', 'S'};
+
+        out << ranks[card.rank] << suits[card.suit];
+        return out;
+    }
+
+    int value() const
+    {
+        static constexpr std::array rankValues{11, 2, 3, 4, 5, 6, 7, 8, 9, 10, 10, 10, 10};
+        return rankValues[rank];
+    }
+};
+
+class Deck
+{
+private:
+    std::array<Card, 52> m_cards{};
+    std::size_t m_nextCardIndex{0};
+
+public:
+    Deck()
+    {
+        std::size_t count{0};
+        for (auto suit : Card::allSuits)
+            for (auto rank : Card::allRanks)
+                m_cards[count++] = Card{rank, suit};
+    }
+
+    void shuffle()
+    {
+        std::shuffle(m_cards.begin(), m_cards.end(), Random::mt);
+        m_nextCardIndex = 0;
+    }
+
+    Card dealCard()
+    {
+        assert(m_nextCardIndex != 52 && "Deck::dealCard ran out of cards");
+        return m_cards[m_nextCardIndex++];
+    }
+};
+
+class Player
+{
+private:
+    int m_score{};
+    int m_aceCount{0};
+
+public:
+    // We'll use a function to add the card to the player's score
+    // Since we now need to count aces
+    void addToScore(Card card)
+    {
+        m_score += card.value();
+        if (card.rank == Card::rank_ace)
+            ++m_aceCount;
+    }
+
+    void consumeAces(int maxScore)
+    {
+        while (m_score > maxScore && m_aceCount > 0)
+        {
+            m_score -= 10;
+            --m_aceCount;
+        }
+    }
+
+    int score() { return m_score; }
+};
 
 namespace Settings
 {
