+++ Jongeun/BlackJack/BlackJack/BlackJack.vcxproj
@@ -130,12 +130,6 @@
   <ItemGroup>
     <ClCompile Include="BlackJack.cpp" />
   </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="Deck.h" />
-    <ClInclude Include="Player.h" />
-    <ClInclude Include="Random.h" />
-    <ClInclude Include="Card.h" />
-  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
