--- Jongeun/BlackJack/Puzzle15/x64/Debug/Puzzle15.exe.recipe
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project>
-  <ProjectOutputs>
-    <ProjectOutput>
-      <FullPath>C:\Users\user\source\repos\SKKUREAL\miniGame\Jongeun\BlackJack\x64\Debug\Puzzle15.exe</FullPath>
-    </ProjectOutput>
-  </ProjectOutputs>
-  <ContentFiles />
-  <SatelliteDlls />
-  <NonRecipeFileRefs />
-</Project>
