--- Jongeun/BlackJack/Puzzle15/x64/Debug/Puzzle15.log
@@ -1,2 +1,0 @@
-﻿  Puzzle15.cpp
-  Puzzle15.vcxproj -> C:\Users\user\source\repos\SKKUREAL\miniGame\Jongeun\BlackJack\x64\Debug\Puzzle15.exe
