--- Jongeun/BlackJack/Puzzle15/x64/Debug/Puzzle15.tlog/Cl.items.tlog
@@ -1,1 +1,0 @@
-C:\Users\user\source\repos\SKKUREAL\miniGame\Jongeun\BlackJack\Puzzle15\Puzzle15.cpp;C:\Users\user\source\repos\SKKUREAL\miniGame\Jongeun\BlackJack\Puzzle15\x64\Debug\Puzzle15.obj
