--- Jongeun/BlackJack/Puzzle15/x64/Debug/Puzzle15.tlog/Puzzle15.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.38.33130:TargetPlatformVersion=10.0.22621.0:
-Debug|x64|C:\Users\user\source\repos\SKKUREAL\miniGame\Jongeun\BlackJack\|
