--- Jongeun/BlackJack/Puzzle15/x64/Debug/Puzzle15.tlog/link.secondary.1.tlog
@@ -1,2 +1,0 @@
-^C:\USERS\USER\SOURCE\REPOS\SKKUREAL\MINIGAME\JONGEUN\BLACKJACK\PUZZLE15\X64\DEBUG\PUZZLE15.OBJ
-C:\Users\user\source\repos\SKKUREAL\miniGame\Jongeun\BlackJack\Puzzle15\x64\Debug\Puzzle15.ilk
