--- Jongeun/BlackJack/Test/x64/Debug/Test.log
@@ -1,5 +1,0 @@
-﻿  test.cpp
-C:\Users\user\source\repos\SKKUREAL\miniGame\Jongeun\BlackJack\Test\test.cpp(20,10): error C2440: 'initializing': cannot convert from 'const int' to 'int &'
-  C:\Users\user\source\repos\SKKUREAL\miniGame\Jongeun\BlackJack\Test\test.cpp(20,10):
-  Conversion loses qualifiers
-
