--- Jongeun/BlackJack/Test/x64/Debug/Test.tlog/CL.command.1.tlog
@@ -1,1 +1,0 @@
-ÿþ
