--- Jongeun/BlackJack/Test/x64/Debug/Test.tlog/link.secondary.1.tlog
@@ -1,2 +1,0 @@
-^C:\USERS\USER\SOURCE\REPOS\SKKUREAL\MINIGAME\JONGEUN\BLACKJACK\TEST\X64\DEBUG\TEST.OBJ
-C:\Users\user\source\repos\SKKUREAL\miniGame\Jongeun\BlackJack\Test\x64\Debug\Test.ilk
